+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 * Add support for executing commands via `hcloud server ssh <server> <command>`
 * Make overriding context via `HCLOUD_CONTEXT` work
+* Add support for JSON and Go template output
 
 ## v1.11.0